+++ packages/components/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- **mi-floor-selector** is a new component for showing a MapsIndoors  Floor Selector.
+- **mi-floor-selector** is a new component for showing a MapsIndoors Floor Selector.
 - - **mi-my-position** is a new component for showing a MapsIndoors My Position button.
 
 ## [13.6.0] - 2023-07-06