+++ packages/components/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [13.7.0] - 2023-07-06
+
+### Added
+
+- **mi-floor-selector** is a new component for showing a MapsIndoors Floor Selector.
+- - **mi-my-position** is a new component for showing a MapsIndoors My Position button.
+
+## [13.6.0] - 2023-07-06
+
+### Added
+
+- **mi-list-item-location** implements `showExternalId` property.
+
 ## [13.5.0] - 2023-07-03
 
 ### Added